+++ package.json
@@ -12,6 +12,8 @@
        },
        "dependencies": {
               "@phosphor-icons/react": "^2.1.10",
+              "@stripe/react-stripe-js": "^3.8.0",
+              "@stripe/stripe-js": "^7.6.1",
               "@tailwindcss/vite": "^4.1.11",
               "autoprefixer": "^10.4.21",
               "bootstrap-icons": "^1.13.1",
