+++ src/features/pages/returnForm/Return.tsx
@@ -1,0 +1,55 @@
+'use client';
+
+import React, { useCallback, useState, useEffect } from "react";
+import {loadStripe} from '@stripe/stripe-js';
+import {
+  EmbeddedCheckoutProvider,
+  EmbeddedCheckout
+} from '@stripe/react-stripe-js';
+import {
+  BrowserRouter,
+  Route,
+  Routes,
+  Navigate
+} from "react-router-dom";
+const stripePromise = loadStripe("pk_test_51RlnqVCRemnaR0EMcYPAKo8USAa7Qhyx8drdhnG2XG4KOp8PrGqOu9D61azLwFxfiaAyOWLcCRIY93W8Z1DXsqAF00z6jI9LR0");
+
+const ReturnForm = () => {
+  const [status, setStatus] = useState(null);
+  const [customerEmail, setCustomerEmail] = useState('');
+
+  useEffect(() => {
+    const queryString = window.location.search;
+    const urlParams = new URLSearchParams(queryString);
+    const sessionId = urlParams.get('session_id');
+
+    fetch(`/session-status?session_id=${sessionId}`)
+      .then((res) => res.json())
+      .then((data) => {
+        setStatus(data.status);
+        setCustomerEmail(data.customer_email);
+      });
+  }, []);
+
+  if (status === 'open') {
+    return (
+      <Navigate to="/checkout" />
+    )
+  }
+
+  if (status === 'complete') {
+    return (
+      <section id="success">
+        <p>
+          We appreciate your business! A confirmation email will be sent to {customerEmail}.
+
+          If you have any questions, please email <a href="mailto:orders@example.com">orders@example.com</a>.
+        </p>
+      </section>
+    )
+  }
+
+  return null;
+}
+
+export default ReturnForm;
